+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 <p>
 <b>Version 1.1.76</b> - August 2020<br/>
+- NEW: The scene controller node can now save the current group address value via a msg input. See here: https://github.com/Supergiovane/node-red-contrib-knx-ultimate/wiki/SceneController-MSG-to-the-node
+</p>
+<p>
+<b>Version 1.1.76</b> - August 2020<br/>
 - Fixed the ETS XML logger. Sometimes the files could contains an invalid CEMI telegram that was not properly discarded and that prevent ETS to load the log. Thanks @Namakemono93.<br/>
 - Fixed a visual glitch in the KNX-Node config window (Telegram value format) for nodered 1.1.0 and above.
 </p>